login function added

+++ app/Http/Controllers/TasksController.php
@@ -17,11 +17,18 @@ class TasksController extends Controller
      */
     public function index()
     {
+        $data = [];
+        if(\Auth::check()){
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            $data = [
+                'user'=>$user,
+                'tasks' => $tasks,
+                ];
+        }
         $tasks = Task::all();
         
-        return view('tasks.index',[
-            'tasks' => $tasks,
-        ]);
+        return view('tasks.index',$data);
     }
 
     /**
@@ -53,12 +60,18 @@ class TasksController extends Controller
             'content' => 'required|max:255',
             'status'  => 'required',
             ]);
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status  = $request->status;
-        $task->save();
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+            ]);
+        return redirect('tasks');
+        // $task = new Task;
+        // $task->content = $request->content;
+        // $task->status  = $request->status;
+        // $task->user_id = 
+        // $task->save();
         
-        return redirect('/');
+        // return redirect('/');
     }
 
     /**
@@ -71,9 +84,13 @@ class TasksController extends Controller
     {
         $task = Task::find($id);
         
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+        if(\Auth::user()->id === $task->user_id){
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+        }else{
+            return redirect('/');
+        }
     }
 
     /**
@@ -85,9 +102,14 @@ class TasksController extends Controller
     public function edit($id)
     {
         $task = Task::find($id);
-        return view('tasks.edit',[
+       
+        if(\Auth::user()->id === $task->user_id){
+             return view('tasks.edit',[
             'task' => $task,
             ]);
+        }else{
+            return redirect('/');
+        }
     }
 
     /**
@@ -105,11 +127,18 @@ class TasksController extends Controller
             ]);
             
         $task = Task::find($id);
-        $task->content = $request->content;
-        $task->status  = $request->status;
-        $task->save();
-        
-        return redirect('/');
+        if(\Auth::user()->id === $task->user_id){ 
+
+            $task->content = $request->content;
+            $task->status  = $request->status;
+            $task->save();
+            
+            return redirect('tasks');
+            
+         }else{
+             
+            return redirect('/');
+        }
     }
 
     /**
@@ -121,8 +150,15 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
         
-        return redirect('/');
+        if(\Auth::user()->id === $task->user_id){ 
+            
+        $task->delete();
+        return redirect('tasks');
+       
+        }else{
+             
+            return redirect('/');
+        }
     }
 }
